+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-Cj5zzeML.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-C4r9d2u0.js";import{_ as a}from"./mermaid.core-C7_-BDQi.js";import"./chunk-5VM5RSS4-Dx7uZyPM.js";import"./chunk-XXDRQBXY-COH7AK9B.js";import"./chunk-POPQ4Y6H-C-V71SRU.js";import"./chunk-F27PBJKO-BByG48CL.js";import"./channel-XNtdqcFw.js";import"./index-EiTDNfOC.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};